requirements.txt
discord 
google-generativeai